- Coupling of HRD locations with HLCD locationIds in HydraulicBoundaryLocation Related to Issue [WTI-330]
git-svn-id: https://repos.deltares.nl/repos/WettelijkToetsInstrumentarium/trunk@2372 5a3db67b-de53-47b5-99c8-a1c30a6650e2

Former-commit-id: f9763d96c595e216869bd4a7163f1e3bfa965282
Former-commit-id: 6f3f794f970f30d9e2eec92a1f8a9832ceff4e11

+++ Ringtoets/HydraRing/src/Ringtoets.HydraRing.IO/HydraulicLocationConfigurationDatabaseContext/HydraulicLocationConfigurationSqLiteDatabaseReader.cs
@@ -59,7 +59,7 @@ namespace Ringtoets.HydraRing.IO.HydraulicLocationConfigurationDatabaseContext
         /// <exception cref="CriticalFileReadException">Thrown when the database query failed.</exception>
         /// <exception cref="InvalidCastException">Thrown when the database returned incorrect values for 
         /// required properties.</exception>
-        public int GetLocationId(int regionId, int hrdLocationId)
+        public int GetLocationId(long regionId, long hrdLocationId)
         {
             var locationIdQuery = HydraulicLocationConfigurationDatabaseQueryBuilder.GetLocationIdQuery();
             var regionParameter = new SQLiteParameter
